% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_ranking.R
\name{compute_gene_ranking}
\alias{compute_gene_ranking}
\title{Compute gene ranking for GSEA based on correlation to cluster centroids}
\usage{
compute_gene_ranking(expr_mat, gene_clusters, cluster1, cluster2)
}
\arguments{
\item{expr_mat}{Numeric matrix of gene expression (genes x samples).}

\item{gene_clusters}{Named vector of cluster assignments for genes. Results of
clustering function}

\item{cluster1}{Cluster number or name for the first cluster.}

\item{cluster2}{Cluster number or name for the second cluster.}
}
\value{
Named numeric vector of gene ranking scores (correlation difference),
sorted in decreasing order.
}
\description{
This function calculates the correlation of each gene's expression profile
within the centroid profiles of two gene clusters, then computes the difference
in correlation to rank genes for GSEA. Positive numbers are for cluster1 and
negative for cluster2
}
\examples{
\dontrun{
ranking <- compute_gene_ranking(normalized, kmeans, 1, 2)
}
}
