% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_enrichment.R
\name{gsea_enrichment}
\alias{gsea_enrichment}
\title{Perform GSEA using clusterProfiler and customizable organism}
\usage{
gsea_enrichment(
  gene_ranking,
  OrgDb,
  keyType = "SYMBOL",
  ont = "BP",
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05
)
}
\arguments{
\item{gene_ranking}{Named numeric vector of gene scores (e.g., log2FC), names are gene IDs.}

\item{OrgDb}{Annotation package, e.g., org.Hs.eg.db or org.Mm.eg.db. Default is org.Hs.eg.db.}

\item{keyType}{The type of gene IDs used in ranking (e.g., "ENSEMBL", "SYMBOL"). Default is SYMBOL.}

\item{ont}{Ontology for GO analysis: "BP", "MF", or "CC". Default is "BP".}

\item{minGSSize}{Minimum size of a gene set. Default is 10.}

\item{maxGSSize}{Maximum size of a gene set. Default is 500.}

\item{pvalueCutoff}{Adjusted p-value cutoff. Default is 0.05.}
}
\value{
A gseaResult object.
}
\description{
This function runs Gene Set Enrichment Analysis (GSEA) using the package clusterProfiler,
assuming the input ranking is a named numeric vector where names are gene identifiers
(e.g., Ensembl IDs). The user can specify the organism database, key type, ontologies,
minimum and maximum size of a gene set and the p-value cutoff.
}
\examples{
\dontrun{
  result <- gsea_enrichment(
    gene_ranking = ranking_kmeans,
    OrgDb = org.Hs.eg.db,
    keyType = "SYMBOL",
    ont = "BP"
  )
}
}
