% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{Plot a Dotplot for a Single GO Enrichment Result}
\usage{
plot_dotplot(enrichment_res, title = "GO Enrichment Dotplot")
}
\arguments{
\item{enrichment_res}{An object of class `enrichResult`, typically the result of `enrichGO()`.}

\item{title}{A character string specifying the plot title.}
}
\value{
A ggplot object representing the dotplot of enrichment results.
}
\description{
This function creates a dotplot from a single GO enrichment result (from `enrichGO`, `enrichKEGG`, etc.).
}
\examples{
\dontrun{
plot_dotplot(go_kmeans[[1]], title = "GO Enrichment - KMeans cluster 1")
}
}
