% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{Plot a Dotplot for a Single GO Enrichment Result}
\usage{
plot_dotplot(enrichment_res, title, showCategory = 10)
}
\arguments{
\item{enrichment_res}{An object of class `enrichResult`, that is the result of `enrichGO()`.}

\item{title}{A character string specifying the plot title.}

\item{showCategory}{Number of enriched term to show in the plot. Default is 10.}
}
\value{
Plots the dotplot of enrichment results.
}
\description{
This function creates a dotplot from a single GO enrichment result.
}
\examples{
\dontrun{
plot_dotplot(go_kmeans[[1]], title = "GO Enrichment - KMeans cluster 1", showCategory = 15)
}
}
