% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap of gene expression grouped by clusters}
\usage{
plot_heatmap(expr_data, clustering_res, scale_data = TRUE)
}
\arguments{
\item{expr_data}{A numeric matrix of raw counts (genes x samples).}

\item{clustering_res}{The result from the `clustering()` function.}

\item{scale_data}{Logical, whether to scale rows before plotting. Default is TRUE.}
}
\value{
The heatmap is plotted.
}
\description{
This function plots a heatmap of gene expression data, ordered by the clustering result.
}
\examples{
\dontrun{
hierarchical <- clustering(normalized, k = 10, method = "hierarchical", distance_method = 'euclidean')
plot_heatmap(normalized, hierarchical)
}
}
